# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Swoosh Commerce API
  version: 1.0.0
  description: Production REST contract. Runtime Swagger UI is available at /api-docs.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refresh_token }
...
security: [{ bearerAuth: [] }]
paths:
...
  /auth/refresh: { post: { security: [{ refreshCookie: [] }], summary: Rotate refresh token, responses: { '200': { description: OK } } } }
  /auth/logout: { post: { security: [{ refreshCookie: [] }], summary: Logout, responses: { '200': { description: OK } } } }
...
  /admin/auth/refresh: { post: { security: [{ refreshCookie: [] }], summary: Admin refresh, responses: { '200': { description: OK } } } }
  /admin/auth/logout: { post: { security: [{ refreshCookie: [] }], summary: Admin logout, responses: { '200': { description: OK } } } }
...